+++ rpc/client_test.go
@@ -23,6 +23,7 @@ import (
 	"time"
 
 	"github.com/cockroachdb/cockroach/proto"
+	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/util"
 	"github.com/cockroachdb/cockroach/util/hlc"
 )
@@ -34,7 +35,7 @@ func init() {
 }
 
 func TestClientHeartbeat(t *testing.T) {
-	tlsConfig, err := LoadTestTLSConfig()
+	tlsConfig, err := security.LoadTestTLSConfig()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -56,7 +57,7 @@ func TestClientHeartbeat(t *testing.T) {
 }
 
 func TestClientNoCache(t *testing.T) {
-	tlsConfig, err := LoadTestTLSConfig()
+	tlsConfig, err := security.LoadTestTLSConfig()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -242,7 +243,7 @@ func (ac *AdvancingClock) UnixNano() int64 {
 // addr. Be sure to close the server when done. Building the server manually
 // like this allows for manual registration of the heartbeat service.
 func createTestServer(serverClock *hlc.Clock, t *testing.T) *Server {
-	tlsConfig, err := LoadTestTLSConfig()
+	tlsConfig, err := security.LoadTestTLSConfig()
 	if err != nil {
 		t.Fatal(err)
 	}
