+++ server/server_test.go
@@ -33,7 +33,7 @@ import (
 	"github.com/cockroachdb/cockroach/client"
 	"github.com/cockroachdb/cockroach/kv"
 	"github.com/cockroachdb/cockroach/proto"
-	"github.com/cockroachdb/cockroach/rpc"
+	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/storage"
 	"github.com/cockroachdb/cockroach/storage/engine"
 	"github.com/cockroachdb/cockroach/util"
@@ -185,7 +185,7 @@ func TestAcceptEncoding(t *testing.T) {
 	s := StartTestServer()
 	defer s.Stop()
 	// We can't use the standard test client. Create our own.
-	tlsConfig := rpc.LoadInsecureClientTLSConfig().Config()
+	tlsConfig := security.LoadInsecureClientTLSConfig().Config()
 	client := &http.Client{
 		Transport: &http.Transport{
 			TLSClientConfig:    tlsConfig,
